backend/functions/transcribe/main.py
import functions_framework
from google.cloud import speech
from google.cloud import firestore
from google.cloud import translate_v2 as translate

db = firestore.Client()
speech_client = speech.SpeechClient()
translate_client = translate.Client()

@functions_framework.http
def transcribe_audio(request):
    """
    HTTP Cloud Function to accept audio data and return transcription/translation.
    Note: For true real-time low latency, a streaming gRPC or WebSocket service 
    is preferred over HTTP Cloud Functions. This is a proof-of-concept for 
    chunk-based processing.
    """
    # TODO: Implement audio handling logic here
    return {"message": "Transcription service ready"}
