backend/functions/transcribe/requirements.txt
functions-framework==3.4.0
google-cloud-speech==2.21.0
google-cloud-firestore==2.11.1
google-cloud-translate==3.11.0
